+++ src/events/handlers/device/common/alerts/lwtStatusAlertHandler.js
@@ -1,14 +1,15 @@
 import logger from '../../../../../common/logger';
 import { WS_DEVICE_ALARM } from '../../../socketConsts';
-import { ALERT_INFO } from '../../../../../common/alertConsts';
+import { levels, types } from '../../../../../common/alertConsts';
 import AlertBuilder from '../../../builders/alertBuilder';
 
 const LwtStatusAlertHandler = (alertProvider, socket) => ({
   process: ({ status, device }) => new Promise((resolve, reject) => {
     logger.log('debug', `lwt alert processor ${JSON.stringify({ device, status })}`);
 
-    const builder = new AlertBuilder(device.gateway, device.deviceId, `${device.description || device.name} is ${status.toUpperCase()}`);
-    builder.setLevel(ALERT_INFO);
+    const builder = new AlertBuilder(device.gateway, device.deviceId, `${device.description || device.name} is ${status.toUpperCase()}`,
+      types.ALERT_TYPE_DEVICE_STATUS);
+    builder.setLevel(levels.ALERT_INFO);
 
     const alert = builder.build();
     alertProvider.add(alert).then(() => {