+++ src/events/handlers/device/common/alerts/powerAlertHandler.spec.js
@@ -1,8 +1,10 @@
 import chai from 'chai';
 import sinon from 'sinon';
-import PowerAlertHandler, { needsNewAlert, makeAlertKey }  from './powerAlertHandler';
+import PowerAlertHandler from './powerAlertHandler';
+import * as alertConsts from '../../../../../common/alertConsts';
 
 import helper from '../../../processor_tests_helper.spec';
+
 helper('./energyAlertProcessor');
 
 chai.should();
@@ -10,10 +12,10 @@ const expect = chai.expect;
 
 
 describe('PowerAlertHandler', () => {
-
- context('Main module', () => {
+  context('Main module', () => {
     let subject;
-    let deviceProvider, alertProvider;
+    let deviceProvider,
+      alertProvider;
     let socket;
 
     beforeEach(() => {
@@ -26,13 +28,49 @@ describe('PowerAlertHandler', () => {
       };
       alertProvider = {
         getLastAlertByKey: () => {},
-        update: () => {}
+        update: () => {},
       };
       socket = {};
       subject = new PowerAlertHandler(deviceProvider, alertProvider, socket);
     });
 
+    it('should create the alert key for the device', (done) => {
+      const stubAlert = sinon.stub(alertConsts, 'alertKey');
 
+      const fakeDevice = {
+        gateway: 'TESTGW',
+        name: 'lamp_test',
+        swVersion: '1.2.3',
+        deviceType: 'Sonoff Pow Module',
+        deviceId: '00:11:22:33:44:55',
+        commands: {
+          power: 'mqtt:cmnd/lamp_test/POWER',
+        },
+        created: new Date(),
+        status: {
+          power: 'on',
+        },
+      };
+
+      sinon.stub(deviceProvider, 'findByDeviceId').returns(Promise.resolve(fakeDevice));
+
+      sinon.stub(alertProvider, 'getLastAlertByKey')
+       .returns(Promise.resolve(null));
+
+      socket.emit = sinon.stub();
+
+      const event = {
+        deviceId: '00:11:22:33:44:55',
+        power: 0,
+      };
+
+      subject.process(event).then(() => {
+        stubAlert.called.should.be.true;
+        stubAlert.calledWith(alertConsts.types.ALERT_TYPE_DEVICE_BROKEN, fakeDevice.gateway, fakeDevice.deviceId).should.be.true;
+        stubAlert.restore();
+        done();
+      });
+    });
 
     it('should create alert when device status is on and there aren\'t previous alerts for the device/gateway', (done) => {
       const event = {
@@ -73,14 +111,81 @@ describe('PowerAlertHandler', () => {
           alertProvider.update
             .calledOnce.should.be.true;
 
-            alertProvider.update
+          alertProvider.update
             .calledWith(sinon.match({
               gateway: 'TESTGW',
               deviceId: '00:11:22:33:44:55',
               message: 'lamp_test could be broken, power is 0 while state is on',
               read: false,
+              open: true,
+              type: 'Device_broken',
               level: 'critical',
-              key: 'alert:energy:TESTGW:00:11:22:33:44:55',
+              key: 'alert:Device_broken:TESTGW:00:11:22:33:44:55',
+            })).should.be.true;
+
+          socket.emit.calledOnce.should.be.true;
+          done();
+        })
+        .catch(err => done(err));
+    });
+
+    it('should update previous alert when device status is on and there is a previous alert', (done) => {
+      const event = {
+        deviceId: '00:11:22:33:44:55',
+        power: 0,
+      };
+
+      const eventDate = new Date();
+
+      sinon.stub(deviceProvider, 'findByDeviceId')
+        .returns(Promise.resolve({
+          gateway: 'TESTGW',
+          name: 'lamp_test',
+          swVersion: '1.2.3',
+          deviceType: 'Sonoff Pow Module',
+          deviceId: '00:11:22:33:44:55',
+          commands: {
+            power: 'mqtt:cmnd/lamp_test/POWER',
+          },
+          created: eventDate,
+          status: {
+            power: 'on',
+          },
+        }));
+
+      sinon.stub(alertProvider, 'getLastAlertByKey')
+        .returns(Promise.resolve({
+          gateway: 'VG59',
+          deviceId: '5C:CF:7F:A0:16:46',
+          message: 'LAMP60 is OFFLINE',
+          date: new Date(),
+          read: false,
+          open: true,
+          level: 'info',
+        }));
+
+      sinon.stub(alertProvider, 'update');
+
+      socket.emit = sinon.stub();
+
+      subject.process(event)
+        .then(() => {
+          deviceProvider.findByDeviceId
+            .called.should.be.true;
+
+          alertProvider.update
+            .calledOnce.should.be.true;
+
+          alertProvider.update
+            .calledWith(sinon.match({},{
+              gateway: 'VG59',
+              deviceId: '5C:CF:7F:A0:16:46',
+              message: 'LAMP60 is OFFLINE',
+              date: new Date(),
+              read: false,
+              open: true,
+              level: 'info',
+              lastUpdate: sinon.match.date,
             })).should.be.true;
 
           socket.emit.calledOnce.should.be.true;
@@ -89,43 +194,4 @@ describe('PowerAlertHandler', () => {
         .catch(err => done(err));
     });
   });
-
-  context('needsNewAlert', () => {
-
-    it('should return true if alert is null', () => {
-      needsNewAlert(null).should.be.true;
-    });
-
-    it('should return true if alert is older than delay', () => {
-      const now = new Date('2017-05-11 10:00:00.000Z');
-      const last = new Date('2017-05-11 09:54:59.000Z')
-
-      needsNewAlert({
-        date: now,
-        lastUpdate: last,
-      }, now, 300).should.be.true;
-    });
-
-    it('should return false if alert isn\'t older than delay', () => {
-      const now = new Date('2017-05-11 10:00:00.000Z');
-      const last = new Date('2017-05-11 09:56:00.000Z');
-
-      needsNewAlert({
-        date: now,
-        lastUpdate: last,
-      }, now, 300).should.be.false;
-    });
-  });
-
-  context('makeAlertKey', () => {
-    it('should create the alert key for the device', () => {
-      makeAlertKey({
-        name: 'lamp1',
-        gateway: 'VG59',
-        swVersion: '3.2.14',
-        deviceType: 'Sonoff Pow Module',
-        deviceId: '5C:CF:7F:A0:16:46',
-      }).should.be.eq('alert:energy:VG59:5C:CF:7F:A0:16:46');
-    })
-  })
 });